defconfig: Update default IO scheduler and CPU frequency governor
Set default IO scheduler to 'noop' as it is designed for non-rotational
media and should be more appropriate.  We can do benchmarks later, but
I feel like this is a sane default for now.

Set default CPU frequency governor to 'ondemand'.  Performance is a waste
of battery, and AOSP usually defaults to ondemand anyways.

Change-Id: Icebd7930d568c6481795830fe0d5527ba229b182